update test for quiz
