+++ front/src/__test__/reducks/quizzes/actions.spec.ts
@@ -1,30 +1,35 @@
 import {
-  FETCH_QUIZZES,
-  DELETE_QUIZ,
   fetchQuizzesAction,
   deleteQuizAction,
 } from "../../../reducks/quizzes/actions";
 
 describe("actions.jsのテスト", () => {
   it("fetchQuizzesActionのテスト", () => {
-    const dummy = [{ id: 1, title: "neko"}, { id: 2, title: "cat" }];
+    const dummy = [
+      {
+        id: 1, title: "neko", created_at: "2021-01-01", updated_at: "2021-02-01"
+      },
+      {
+        id: 2, title: "cat", created_at: "2021-01-01", updated_at: "2021-02-01"
+      }
+    ];
     const action = fetchQuizzesAction(dummy);
 
     expect(action).toStrictEqual(
       {
-        type: FETCH_QUIZZES,
+        type: "FETCH_QUIZZES",
         payload: dummy,
       }
     );
   });
   
   it("deleteQuizActionのテスト", () => {
-    const dummy = { id: 1, title: "neko" };
+    const dummy = [{ id: 1, title: "neko", created_at: "2021-01-01", updated_at: "2021-02-01" }];
     const action = deleteQuizAction(dummy);
 
     expect(action).toStrictEqual(
       {
-        type: DELETE_QUIZ,
+        type: "DELETE_QUIZ",
         payload: dummy,
       }
     );
